// src/payload/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  caption?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
// ...
export interface NavCollection {
  id: string;
  title: string;
  navGroup?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: string | Page;
          } | null;
          url?: string | null;
          label: string;
          toltip?:
            | {
                toltipType?: ('custom' | 'new') | null;
                customMessage?: string | null;
                id?: string | null;
              }[]
            | null;
          icon?:
            | {
                iconType?:
                  | (
                      | 'facebook'
                      | 'whatsapp'
                      | 'youtube'
                      | 'instagram'
                      | 'linkedin'
                      | 'discord'
                      | 'telegram'
                      | 'twitter'
                      | 'github'
                      | 'custom'
                    )
                  | null;
                iconImage?: string | Media | null;
                id?: string | null;
              }[]
            | null;
        };
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Header {
  id: string;
  brandLogo?: string | Media | null;
  navItems?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?: {
            relationTo: 'pages';
            value: string | Page;
          } | null;
          url?: string | null;
          label: string;
          toltip?:
            | {
                toltipType?: ('custom' | 'new') | null;
                customMessage?: string | null;
                id?: string | null;
              }[]
            | null;
          icon?:
            | {
                iconType?:
                  | (
                      | 'facebook'
                      | 'whatsapp'
                      | 'youtube'
                      | 'instagram'
                      | 'linkedin'
                      | 'discord'
                      | 'telegram'
                      | 'twitter'
                      | 'github'
                      | 'custom'
                    )
                  | null;
                iconImage?: string | Media | null;
                id?: string | null;
              }[]
            | null;
        };
        navCollections?: (string | null) | NavCollection;
        id?: string | null;
      }[]
    | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}